// frontend/src/types/website.ts
export interface Service {
    id: string;
    title: string;
    description: string;
    iconUrl: string;
    order: number;
}